style(Hero.module.css): Increase background-size to 55% and background-position to 105% 25% refactor(ArtistCard.tsx): Move ArtistCard component from app/components/artists/ArtistCard.tsx to app/routes/($locale).artists.tsx refactor(convert-to-array.ts): Export convertToEmbedArray function from app/data/convert-to-array.ts refactor(($locale)._index.tsx): Import convertToEmbedArray function from app/data/convert-to-array.ts refactor(($locale).artists.tsx): Move ArtistCard component from app/components/artists/ArtistCard.tsx to app/routes/($locale).artists.tsx
