+++ app/data/convert-to-array.ts
@@ -1,4 +1,4 @@
 // https://community.shopify.com/c/technical-q-a/graphql-and-product-metafields-how-to-access-array-of-products/m-p/1611955
-const convertToEmbedArray = (sampleArr: string): string[] => {
+export const convertToEmbedArray = (sampleArr: string): string[] => {
   return sampleArr?.replace(/\[|\]/g, '').replace(/"/g, '').split(',') ?? [];
 };
